refactor: code refactor and update docs;
