+++ docs/references.bib
@@ -1,0 +1,66 @@
+@article{zhang2017airquality,
+title={Cautionary tales on air-quality improvement in Beijing},
+author={Shuyi Zhang and Bin Guo and Anlan Dong and Jing He and Ziping Xu and Song Xi Chen},
+journal={Proceedings of the Royal Society A: Mathematical, Physical and Engineering Sciences},
+year={2017},
+volume={473},
+url={https://api.semanticscholar.org/CorpusID:37683936}
+}
+
+@misc{trindade2015electricity,
+author       = {Trindade,Artur},
+title        = {{ElectricityLoadDiagrams20112014}},
+year         = {2015},
+howpublished = {UCI Machine Learning Repository},
+note         = {{DOI}: https://doi.org/10.24432/C58C86}
+}
+
+@article{silva2012physionet,
+title={Predicting in-hospital mortality of icu patients: The physionet/computing in cardiology challenge 2012},
+author={Silva, Ikaro and Moody, George and Scott, Daniel J and Celi, Leo A and Mark, Roger G},
+journal={Computing in cardiology},
+volume={39},
+pages={245},
+year={2012},
+publisher={NIH Public Access}
+}
+
+@article{reyna2019physionet,
+title={Early prediction of sepsis from clinical data: the PhysioNet\/Computing in Cardiology Challenge 2019.},
+author={Reyna, MA and Josef, C and Jeter, R and Shashikumar, SP and Westover, MB and  Nemati, S and Clifford, GD and  Sharma, A},
+journal={Critical Care Medicine},
+volume={48},
+number={2},
+pages={210--217},
+year={2019}
+}
+
+@misc{dau2018ucr,
+title = {The UCR Time Series Classification Archive},
+author = {Dau, Hoang Anh and Keogh, Eamonn and Kamgar, Kaveh and Yeh, Chin-Chia Michael and Zhu, Yan
+          and Gharghabi, Shaghayegh and Ratanamahatana, Chotirat Ann and Yanping and Hu, Bing
+          and Begum, Nurjahan and Bagnall, Anthony and Mueen, Abdullah and Batista, Gustavo, and Hexagon-ML},
+year = {2018},
+month = {October},
+note = {\url{https://www.cs.ucr.edu/~eamonn/time_series_data_2018/}}
+}
+
+@misc{bagnall2018uea,
+title={The UEA multivariate time series classification archive, 2018},
+author={Anthony Bagnall and Hoang Anh Dau and Jason Lines and Michael Flynn and James Large and Aaron Bostrom and Paul Southam and Eamonn Keogh},
+year={2018},
+eprint={1811.00075},
+archivePrefix={arXiv},
+primaryClass={cs.LG}
+}
+
+@misc{grgicevic2023ais,
+author       = {Luka Grgičević},
+title        = {AIS data},
+month        = jun,
+year         = 2023,
+publisher    = {Zenodo},
+version      = {1.0.1},
+doi          = {10.5281/zenodo.8064564},
+url          = {https://doi.org/10.5281/zenodo.8064564}
+}
