+++ docs/references.rst
@@ -1,0 +1,5 @@
+References
+==========
+
+.. bibliography::
+   :style: unsrt
