--- docs/tsdb.data_loading_funcs.rst
@@ -1,68 +1,0 @@
-tsdb.data\_loading\_funcs package
-=================================
-
-Submodules
-----------
-
-tsdb.data\_loading\_funcs.beijing\_multisite\_air\_quality module
------------------------------------------------------------------
-
-.. automodule:: tsdb.data_loading_funcs.beijing_multisite_air_quality
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :inherited-members:
-
-tsdb.data\_loading\_funcs.electricity\_load\_diagrams module
-------------------------------------------------------------
-
-.. automodule:: tsdb.data_loading_funcs.electricity_load_diagrams
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :inherited-members:
-
-tsdb.data\_loading\_funcs.physionet\_2012 module
-------------------------------------------------
-
-.. automodule:: tsdb.data_loading_funcs.physionet_2012
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :inherited-members:
-
-tsdb.data\_loading\_funcs.vessel_AIS module
-------------------------------------------------
-
-.. automodule:: tsdb.data_loading_funcs.vessel_AIS
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :inherited-members:
-
-tsdb.data\_loading\_funcs.physionet\_2019 module
-------------------------------------------------
-
-.. automodule:: tsdb.data_loading_funcs.physionet_2019
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :inherited-members:
-
-tsdb.data\_loading\_funcs.ucr\_uea\_datasets module
----------------------------------------------------
-
-.. automodule:: tsdb.data_loading_funcs.ucr_uea_datasets
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :inherited-members:
-
-Module contents
----------------
-
-.. automodule:: tsdb.data_loading_funcs
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :inherited-members:
