+++ docs/tsdb.rst
@@ -1,39 +1,6 @@
 tsdb package
 ============
 
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   tsdb.data_loading_funcs
-   tsdb.tests
-
-Submodules
-----------
-
-tsdb.data\_processing module
-----------------------------
-
-.. automodule:: tsdb.data_processing
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :inherited-members:
-
-tsdb.database module
---------------------
-
-.. automodule:: tsdb.database
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :inherited-members:
-
-Module contents
----------------
-
 .. automodule:: tsdb
    :members:
    :undoc-members:
